Change non-functioning test to example
