+++ src/gcp/mod.rs
@@ -311,80 +311,4 @@ mod test {
         assert_eq!(expected, hash);
         assert_eq!("scope1 scope2 scope3", scopes);
     }
-
-    //#[test]
-    #[allow(dead_code)]
-    fn test_token_retrieval() {
-        const SERVICE_KEY: &str = include_str!("../../tests/svc_key.json");
-
-        use bytes::BufMut;
-
-        let acct_info = ServiceAccountInfo::deserialize(SERVICE_KEY).unwrap();
-        let acct_access = ServiceAccountAccess::new(acct_info).unwrap();
-
-        let token = match acct_access
-            .get_token(&["https://www.googleapis.com/auth/pubsub"])
-            .unwrap()
-        {
-            TokenOrRequest::Request {
-                request,
-                scope_hash,
-                ..
-            } => {
-                let client = reqwest::Client::new();
-
-                let (parts, body) = request.into_parts();
-                let uri = parts.uri.to_string();
-
-                let builder = match parts.method {
-                    http::Method::GET => client.get(&uri),
-                    http::Method::POST => client.post(&uri),
-                    http::Method::DELETE => client.delete(&uri),
-                    http::Method::PUT => client.put(&uri),
-                    method => unimplemented!("{} not implemented", method),
-                };
-
-                let request = builder.headers(parts.headers).body(body).build().unwrap();
-
-                let mut response = client.execute(request).unwrap();
-
-                let mut writer = bytes::BytesMut::with_capacity(
-                    response.content_length().unwrap_or(1024) as usize,
-                )
-                .writer();
-                response.copy_to(&mut writer).unwrap();
-                let buffer = writer.into_inner();
-
-                let mut builder = http::Response::builder();
-
-                builder
-                    .status(response.status())
-                    .version(response.version());
-
-                let headers = builder.headers_mut().unwrap();
-
-                headers.extend(
-                    response
-                        .headers()
-                        .into_iter()
-                        .map(|(k, v)| (k.clone(), v.clone())),
-                );
-
-                let response = builder.body(buffer.freeze()).unwrap();
-
-                acct_access
-                    .parse_token_response(scope_hash, response)
-                    .unwrap()
-            }
-            _ => unreachable!(),
-        };
-
-        match acct_access
-            .get_token(&["https://www.googleapis.com/auth/pubsub"])
-            .unwrap()
-        {
-            TokenOrRequest::Token(tk) => assert_eq!(tk, token),
-            _ => unreachable!(),
-        }
-    }
 }
